清理过期图片逻辑

+++ ZSCY_Win10/Util/StartPage/StartPageHelp.cs
@@ -119,8 +119,11 @@ namespace ZSCY_Win10.Util.StartPage
                         }
                         else
                         {
-                            File.Delete(item.Url);
-                            DeleteDatabase(item);
+                            if (DateTime.Now > Convert.ToDateTime(item.StartTime))
+                            {
+                                File.Delete(item.Url);
+                                DeleteDatabase(item);
+                            }
                         }
 #endif
                     }
